# apps/rust-server/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InstantML Rust API",
    "description": "Current Rust/ClickHouse API. See docs/architecture/current-api.md for examples and operational notes.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/demo/reset": {
      "post": {
        "tags": [
          "runs"
        ],
        "operationId": "reset_demo",
        "responses": {
          "200": {
            "description": "Reset deterministic demo project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObjectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing ingest scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "browserSession": []
          }
        ]
      }
    },
...
    "/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "operationId": "list_runs",
        "parameters": [
          {
            "name": "project",
            "in": "query",
            "description": "Filter by project name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by run status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (1..=1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of runs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "browserSession": []
          }
        ]
      },
      "post": {
        "tags": [
          "runs"
        ],
        "operationId": "create_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "browserSession": []
          }
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "operationId": "get_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Run UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "browserSession": []
          }
        ]
      },
      "patch": {
        "tags": [
          "runs"
        ],
        "operationId": "update_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Run UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "browserSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "type": "object",
        "properties": {
          "config": {
            "type": [
              "object",
              "null"
            ]
          },
          "metadata": {
            "type": [
              "object",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "project": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "JsonObjectResponse": {
        "type": "object",
        "description": "Free-form key/value response payload used by handlers that build dynamic\nJSON objects in the store layer (overview, usage, exports, etc.). Modeled\nas an opaque object so the OpenAPI spec is honest about the shape without\nduplicating the dynamic structure here.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
      "RunEnvelope": {
        "type": "object",
        "required": [
          "run"
        ],
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunRow"
          }
        }
      },
      "RunRow": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "project_id",
          "project",
          "name",
          "status",
          "config",
          "tags",
          "metadata",
          "created_at",
          "started_at"
        ],
        "properties": {
          "config": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "finished_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "project": {
            "type": "string"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RunsEnvelope": {
        "type": "object",
        "required": [
          "runs"
        ],
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunRow"
            }
          }
        }
      },
...
      "UpdateRunRequest": {
        "type": "object",
        "properties": {
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "InstantML SDK API key sent as Authorization: Bearer instantml_..."
      },
      "bootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-InstantML-Bootstrap-Token",
        "description": "Operator-only bootstrap token for initial users, orgs, and admin key paths"
      },
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "instantml_session",
        "description": "HttpOnly browser session cookie issued by /api/auth/dev/google or /api/auth/clerk"
      }
    }
  },
  "security": [
    {
      "bearerApiKey": []
    },
    {
      "browserSession": []
    }
  ],
...
}